refactor(RVCV-163): 리뷰 속성 조회 api 위치 변경(shopadmin -> reviewproperty)
