+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/reviewproperty/application/usecase/ReviewPropertyUseCase.java
@@ -1,0 +1,8 @@
+package com.romanticpipe.reviewcanvas.domain.reviewproperty.application.usecase;
+
+import com.romanticpipe.reviewcanvas.domain.reviewproperty.application.usecase.response.GetReviewPropertyResponse;
+
+public interface ReviewPropertyUseCase {
+
+	GetReviewPropertyResponse getAllReviewProperty(Integer adminId);
+}
