+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/reviewproperty/application/usecase/ReviewPropertyUseCaseImpl.java
@@ -1,0 +1,11 @@
+package com.romanticpipe.reviewcanvas.domain.reviewproperty.application.usecase;
+
+import com.romanticpipe.reviewcanvas.domain.reviewproperty.application.usecase.response.GetReviewPropertyResponse;
+
+public class ReviewPropertyUseCaseImpl implements ReviewPropertyUseCase {
+
+	@Override
+	public GetReviewPropertyResponse getAllReviewProperty(Integer adminId) {
+		return null;
+	}
+}
