+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/shopadmin/application/usecase/ShopAdminUseCase.java
@@ -1,6 +1,5 @@
 package com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase;
 
-import com.romanticpipe.reviewcanvas.domain.reviewproperty.application.usecase.response.GetReviewPropertyResponse;
 import com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.request.SignUpRequest;
 
 public interface ShopAdminUseCase {
@@ -9,5 +8,4 @@ public interface ShopAdminUseCase {
 
 	boolean emailCheck(String email);
 
-	GetReviewPropertyResponse getAllReviewProperty(Integer adminId);
 }
